Experiment with early return atop the existing GooseLang; delete stuff
that depends on old `typing.v`. Want to isolate the part of the type
system that is used for defining the semantics of Go (and is thus
trusted) v.s. the part that is convenient for proof.  I think most of
the typing rules are only important for assuming a well-typed program in
the refinement proof. The definitions that are useful for that theorem
should not infect the entire program logic.

It's in general difficult to point to the entirety of the trusted formal
definition of Go programs (e.g. it includes various lib/*/impl.v's).
Cleaning that up a bit to make faster progress with Goose upgrades.